requirements.txt
Flask==3.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
itsdangerous==2.1.2
click==8.1.7
requests==2.31.0
Flask-SQLAlchemy==3.1.1
celery==5.3.6
redis==5.0.1
Flask-Login==0.6.3
Flask-Security==3.0.0
python-dotenv==1.0.0
flake8==6.1.0
black==23.12.1
pytest==7.4.3
pymysql==1.1.0